[feature] Function first draft of matlab_sim_driver
Had to restamp the pose transforms in order to avoid the TF_REPEATED_DATA warning. Its hard to understand why that is happening when we are only republishing the messages we receive from simulink with their original stamp which should almost always make sure that no two messages have the same stamp.