[fix] Changed to rotation matrix based computation of angular velocity
It is possible that the scalar term of q might flip between negative and positive values if for some reason vicon doesn't enforce the canonical map. If that happens q_dot will just blow up and lead to faulty angular velocity estimates.
Thus, it is better to use the rotation matrix which we know will only evolve smoothly in SO(3).