[features, refactors] JIT compatible quaternion algebra
Yet to implement getting the rotation matrix from quaternions with JIT, that should considerably speed up the control loop.
Made the ACTIVE COILS and the ACTIVE DRIVERS variables hidden from the derived classes since they should not be accidently changed. Otherwise we may get some errors. This is the reason that going forward, the allocation matrix computation should be a feature of the base class and not of the derived classes. The JIT compatibility is to ensure that live calculation of forces and torques with ZOH assumption on the currents can be performed in real time with the simulator. However, this is low priority as of this commit.