[feature] Y Controller Tested, Observation note taking in experiment analysis
The experiment analysis node now opens a text editor to note down observations about a certain experiment when run. It also logs every time the node was run so that observations can also be timestamped in case that is of siginificance, for later realizations for example. The Y controller was tested and the observations have been noted down.