[major fix] Bug in individual magnet allocatoin
I found a bug in my individual magnet allocation matrix computation code. The inertial frame forces were not being transformed to the body fixed M frame. It was missing a matrix multiplication which has now been added.
It was an easy mistake to make because the cross product with a z aligned t_MG_M (displacement of magnet w.r.t COM frame M expressed in M) still led to a zero row for local Tau_z's mapping to the currents. But the components for the other axes will still turn out to be incorrect.
It seems like this did not influence my expriments so far that much. The code for the plotting library seems to be fine since it was directly computing the actual torques in the local frame anyways, but I rewrote its code for better readability and made the transformation of inertial frame forces to M more explicit.