[fix] roslaunch escalating on SIGKILL with RT priority
For some reason rosnode kill was still working despite launching the node with SCHED_FIFO since the node checks for shutdown requests but roslaunch was unable to stop the node (seemingly) despite the shutdown request messages and the stop service clearly being run. So resetting to the default scheduler was added and that fixed the issue.